+++ packages/tsurlfilter/src/index.ts
@@ -27,4 +27,5 @@ export * from './utils/string-utils';
 export { CosmeticRuleMarker } from './rules/cosmetic-rule-marker';
 export { CosmeticRuleParser } from './rules/cosmetic-rule-parser';
 export { RuleSyntaxUtils } from './utils/rule-syntax-utils';
+export { HTTPMethod } from './modifiers/method-modifier';
 export { NETWORK_RULE_OPTIONS, OPTIONS_DELIMITER } from './rules/network-rule-options';