+++ packages/tswebextension/src/lib/mv2/background/web-request-api.ts
@@ -307,6 +307,7 @@ export class WebRequestApi {
             frameUrl: referrerUrl,
             requestType,
             frameRule: tabsApi.getTabFrameRule(tabId),
+            method,
         });
 
         if (!result) {