+++ packages/tsurlfilter/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [1.0.77]: https://github.com/AdguardTeam/tsurlfilter/compare/tsurlfilter-v1.0.76...tsurlfilter-v1.0.77 -->
 
+[Unreleased]
+
 
 ## [2.1.0] - 2023-06-14
 
@@ -27,6 +29,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `$redirect` rules are higher in priority than blocking rules.
 - Exceptions are higher in priority than `$redirect` and blocking rules.
 - `$important` is higher than all others.
+
+## [Unreleased]
+
+### Added
+
+- new `$method` modifier to match requests by HTTP method
+
 
 ## [2.0.7] - 2023-06-14